@@ -7,9 +7,7 @@ import 'package:weather_app/repositories/repositories.dart';
 class WeatherBloc extends IsolateBloc<WeatherEvent, WeatherState> {
   final WeatherRepository weatherRepository;
 
-  WeatherBloc({@required this.weatherRepository})
-      : assert(weatherRepository != null),
-        super(WeatherInitial());
+  WeatherBloc({required this.weatherRepository}) : super(WeatherInitial());
 
   @override
   void onEventReceived(WeatherEvent event) {
@@ -48,7 +46,7 @@ abstract class WeatherEvent extends Equatable {
 class WeatherRequested extends WeatherEvent {
   final String city;
 
-  const WeatherRequested({@required this.city}) : assert(city != null);
+  const WeatherRequested({required this.city});
 
   @override
   List<Object> get props => [city];
@@ -57,7 +55,7 @@ class WeatherRequested extends WeatherEvent {
 class WeatherRefreshRequested extends WeatherEvent {
   final String city;
 
-  const WeatherRefreshRequested({@required this.city}) : assert(city != null);
+  const WeatherRefreshRequested({required this.city});
 
   @override
   List<Object> get props => [city];
@@ -77,7 +75,7 @@ class WeatherLoadInProgress extends WeatherState {}
 class WeatherLoadSuccess extends WeatherState {
   final Weather weather;
 
-  const WeatherLoadSuccess({@required this.weather}) : assert(weather != null);
+  const WeatherLoadSuccess({required this.weather});
 
   @override
   List<Object> get props => [weather];
