@@ -10,7 +10,7 @@ class WeatherApiClient {
   static const baseUrl = 'https://www.metaweather.com';
   final http.Client httpClient;
 
-  WeatherApiClient({@required this.httpClient}) : assert(httpClient != null);
+  WeatherApiClient({required this.httpClient});
 
   Future<int> getLocationId(String city) async {
     final locationUrl = '$baseUrl/api/location/search/?query=$city';
