@@ -8,8 +8,7 @@ import 'weather_api_client.dart';
 class WeatherRepository {
   final WeatherApiClient weatherApiClient;
 
-  WeatherRepository({@required this.weatherApiClient})
-      : assert(weatherApiClient != null);
+  WeatherRepository({required this.weatherApiClient});
 
   Future<Weather> getWeather(String city) async {
     final int locationId = await weatherApiClient.getLocationId(city);
