@@ -7,11 +7,10 @@ class GradientContainer extends StatelessWidget {
   final MaterialColor color;
 
   const GradientContainer({
-    Key key,
-    @required this.color,
-    @required this.child,
-  })  : assert(color != null, child != null),
-        super(key: key);
+    Key? key,
+    required this.color,
+    required this.child,
+  }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -22,9 +21,9 @@ class GradientContainer extends StatelessWidget {
           end: Alignment.bottomCenter,
           stops: [0.6, 0.8, 1.0],
           colors: [
-            color[700],
-            color[500],
-            color[300],
+            color[700]!,
+            color[500]!,
+            color[300]!,
           ],
         ),
       ),
