@@ -5,9 +5,7 @@ import 'package:meta/meta.dart';
 class LastUpdated extends StatelessWidget {
   final DateTime dateTime;
 
-  LastUpdated({Key key, @required this.dateTime})
-      : assert(dateTime != null),
-        super(key: key);
+  LastUpdated({Key? key, required this.dateTime}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
