@@ -6,9 +6,7 @@ import 'package:weather_app/models/models.dart';
 class WeatherConditions extends StatelessWidget {
   final WeatherCondition condition;
 
-  WeatherConditions({Key key, @required this.condition})
-      : assert(condition != null),
-        super(key: key);
+  WeatherConditions({Key? key, required this.condition}) : super(key: key);
 
   @override
   Widget build(BuildContext context) => _mapConditionToImage(condition);
