@@ -1,6 +1,7 @@
-import 'isolated_bloc_manager.dart';
-import 'platform_channel/isolated_platform_channel_middleware.dart';
-import 'service_events.dart';
+
+import 'package:isolate_bloc/src/common/isolate/isolated_bloc_manager.dart';
+import 'package:isolate_bloc/src/common/isolate/platform_channel/isolated_platform_channel_middleware.dart';
+import 'package:isolate_bloc/src/common/isolate/service_events.dart';
 
 /// Listen for [ServiceEvent]s in Isolate
 class IsolatedConnector {
