@@ -1,7 +1,7 @@
 import 'package:flutter/widgets.dart';
+import 'package:isolate_bloc/src/flutter/isolate_bloc_provider.dart';
 import 'package:provider/provider.dart';
 
-import 'isolate_bloc_provider.dart';
 
 /// {@template multi_bloc_provider}
 /// Merges multiple [IsolateBlocProvider] widgets into one widget tree.
