@@ -1,8 +1,6 @@
 import 'package:isolate_bloc/isolate_bloc.dart';
-import 'package:isolate_bloc/src/common/isolate/isolate_manager/isolate/isolate_manager.dart'
-    as native_isolate;
-import 'package:isolate_bloc/src/common/isolate/isolate_manager/web/isolate_manager.dart'
-    as web_isolate;
+import 'package:isolate_bloc/src/common/isolate/isolate_manager/isolate/isolate_manager.dart';
+import 'package:isolate_bloc/src/common/isolate/isolate_manager/web/isolate_manager.dart';
 
 TestInitializePlatform? _testInitializePlatform;
 
@@ -17,20 +15,22 @@ Future<void> testInitialize(Initializer userInitializer) async {
   );
   return BlocManager.initialize(
     userInitializer,
-    () {
-      switch (_testInitializePlatform!) {
-        case TestInitializePlatform.web:
-          return native_isolate.IsolateManagerImpl.createIsolate;
-        case TestInitializePlatform.native:
-          return web_isolate.WebIsolateManager.createIsolate;
-      }
-    }(),
+    _testFactory,
     [],
   );
 }
 
+IsolateManagerFactory get _testFactory {
+  switch (_testInitializePlatform!) {
+    case TestInitializePlatform.web:
+      return WebIsolateManagerFactory();
+    case TestInitializePlatform.native:
+      return IOIsolateManagerFactory();
+  }
+}
+
 /// Platform which is used to determine which [IsolateManager] implementation to use in tests.
-/// Used to test both web and native backends.
+/// Used to test both web and native backend.
 enum TestInitializePlatform {
   web,
   native,
