@@ -1,6 +1,6 @@
 import 'package:isolate_bloc/isolate_bloc.dart';
 
-/// [IsolateWrapper] for [MockIsolateManager]
+/// [IsolateWrapper] for mock [IsolateManager]
 class MockIsolateWrapper extends IsolateWrapper {
   @override
   void kill() {}
